Fix build error and enhance server side rendering
